Prevent incorrect clipping of text.
When the text bounding box had a negative y0, the text was cropped.
This happened for all vertical alignments except ascender, notably
for baseline, which is needed for aligning text of different sizes.

The new implementation is simpler and more obviously correct, as all
conditions on the sign of x0 and y0 were removed.

Also optimize the size of cached images when x0 or y0 is positive by
saving exactly the bounding box, without blank space on the top left.